review update. moved definitions out of types package into tail.go
