+++ src/cmd/cli/command/service_monitoring.go
@@ -2,13 +2,13 @@ package command
 
 import (
 	"context"
+
 	"github.com/DefangLabs/defang/src/pkg/cli"
 	"github.com/DefangLabs/defang/src/pkg/term"
-	"github.com/DefangLabs/defang/src/pkg/types"
 	defangv1 "github.com/DefangLabs/defang/src/protos/io/defang/v1"
 )
 
-func monitorServiceStatus(ctx context.Context, targetStatus types.ServiceStatus, serviceInfos []*defangv1.ServiceInfo, cancel context.CancelFunc) {
+func monitorServiceStatus(ctx context.Context, targetStatus cli.ServiceStatus, serviceInfos []*defangv1.ServiceInfo, cancel context.CancelFunc) {
 	serviceList := []string{}
 	for _, serviceInfo := range serviceInfos {
 		serviceList = append(serviceList, serviceInfo.Service.Name)
@@ -27,7 +27,7 @@ func monitorServiceStatus(ctx context.Context, targetStatus types.ServiceStatus,
 		if isAllServicesAtSameStatus(targetStatus, *serviceStatus) {
 			cancel()
 			for _, sInfo := range serviceInfos {
-				sInfo.Status = string(types.ServiceStarted)
+				sInfo.Status = string(cli.ServiceStarted)
 			}
 			printEndpoints(serviceInfos)
 			return
@@ -35,10 +35,10 @@ func monitorServiceStatus(ctx context.Context, targetStatus types.ServiceStatus,
 	}
 
 	// server channel closed without having all services completed
-	term.Warnf("service state monitoring terminated without all services reaching desired state: %s", types.ServiceStarted)
+	term.Warnf("service state monitoring terminated without all services reaching desired state: %s", cli.ServiceStarted)
 }
 
-func isAllServicesAtSameStatus(targetStatus types.ServiceStatus, serviceStatuses map[string]string) bool {
+func isAllServicesAtSameStatus(targetStatus cli.ServiceStatus, serviceStatuses map[string]string) bool {
 	allDone := true
 	for _, status := range serviceStatuses {
 		if status != string(targetStatus) {
