+++ src/pkg/cli/tail.go
@@ -32,6 +32,20 @@ var (
 	DoVerbose     = false
 )
 
+type ServiceStatus string
+
+const (
+	ServiceDeploymentStarting   ServiceStatus = "STARTING"
+	ServiceDeploymentInProgress ServiceStatus = "IN_PROGRESS"
+	ServiceStarted              ServiceStatus = "COMPLETED"
+	ServiceStopping             ServiceStatus = "STOPPING"
+	ServiceStopped              ServiceStatus = "STOPPED"
+	ServiceDeactivating         ServiceStatus = "DEACTIVATING"
+	ServiceDeprovisioning       ServiceStatus = "DEPROVISIONING"
+	ServiceFailed               ServiceStatus = "FAILED"
+	ServiceUnknown              ServiceStatus = "UNKNOWN"
+)
+
 type EndLogConditional struct {
 	Service  string
 	Host     string
